move and update form component
